newshour-namedentity-wikipedialink/requirements.txt
requests
mendelai-brat-parser
pandas
smart_open
clams-utils
